Feat: Split up branched tests like logsig and sig tests
