+++ tests/test_branched_sig.py
@@ -21,7 +21,6 @@ import itertools
 import kauri
 import kauri.bck
 import kauri.mkw
-import kauri.mkw
 
 import pysiglib
 from conftest import DEVICES, check_close, skip_no_cuda
@@ -242,27 +241,6 @@ def test_branched_sig_vs_kauri(d, N):
     np.testing.assert_allclose(bsig[1:], ref_reordered, atol=1e-10)
 
 
-@pytest.mark.parametrize("d,N", [(2, 3)])
-def test_branched_sig_concatenation(d, N):
-    """Chen's identity: bsig(X*Y) == branched_sig_combine(bsig(X), bsig(Y))."""
-    pysiglib.prepare_branched_sig(d, N)
-    np.random.seed(99)
-    L1, L2 = 5, 6
-    path1 = np.cumsum(np.random.randn(L1, d) * 0.1, axis=0)
-    # path2 must start where path1 ends
-    path2_increments = np.random.randn(L2 - 1, d) * 0.1
-    path2 = np.vstack([path1[-1:], path1[-1] + np.cumsum(path2_increments, axis=0)])
-
-    bsig1 = pysiglib.branched_sig(path1, N)
-    bsig2 = pysiglib.branched_sig(path2, N)
-    combined = pysiglib.branched_sig_combine(bsig1, bsig2, d, N)
-
-    full_path = np.vstack([path1, path2[1:]])
-    bsig_full = pysiglib.branched_sig(full_path, N)
-
-    np.testing.assert_allclose(combined, bsig_full, atol=1e-10)
-
-
 @pytest.mark.parametrize("d,N", [(2, 3), (3, 2)])
 def test_branched_sig_batch(d, N):
     pysiglib.prepare_branched_sig(d, N)
@@ -380,27 +358,6 @@ def test_branched_sig_time_aug_batch(d, N):
     for i in range(B):
         single = pysiglib.branched_sig(paths[i], N, time_aug=True)
         np.testing.assert_allclose(batch_result[i], single, atol=1e-12)
-
-
-@pytest.mark.parametrize("d,N", [(2, 3)])
-def test_branched_sig_lead_lag_concatenation(d, N):
-    """Chen's identity should hold with lead_lag enabled."""
-    aug_dim = 2 * d
-    pysiglib.prepare_branched_sig(aug_dim, N)
-    np.random.seed(204)
-    L1, L2 = 5, 6
-    path1 = np.cumsum(np.random.randn(L1, d) * 0.1, axis=0)
-    path2_inc = np.random.randn(L2 - 1, d) * 0.1
-    path2 = np.vstack([path1[-1:], path1[-1] + np.cumsum(path2_inc, axis=0)])
-
-    bsig1 = pysiglib.branched_sig(path1, N, lead_lag=True)
-    bsig2 = pysiglib.branched_sig(path2, N, lead_lag=True)
-    combined = pysiglib.branched_sig_combine(bsig1, bsig2, aug_dim, N)
-
-    full_path = np.vstack([path1, path2[1:]])
-    bsig_full = pysiglib.branched_sig(full_path, N, lead_lag=True)
-
-    np.testing.assert_allclose(combined, bsig_full, atol=1e-10)
 
 
 # ---------------------------------------------------------------------------
@@ -571,6 +528,26 @@ def test_branched_sig_backprop_torch_api(d, N):
 
 
 @pytest.mark.parametrize("d,N", [(2, 3)])
+def test_planar_branched_sig_backprop_torch_api(d, N):
+    """torch_api planar branched_sig backward produces correct gradients."""
+    pysiglib.prepare_branched_sig(d, N, planar=True)
+    np.random.seed(4021)
+    path = torch.tensor(np.cumsum(np.random.randn(5, d) * 0.1, axis=0),
+                        dtype=torch.float64, requires_grad=True)
+
+    bsig = pysiglib.torch_api.branched_sig(path, N, planar=True)
+    loss = bsig.sum()
+    loss.backward()
+    grad_torch = path.grad.clone()
+
+    grad_manual = pysiglib.branched_sig_backprop(
+        path.detach(), bsig.detach(),
+        torch.ones_like(bsig), N, planar=True)
+
+    check_close(grad_torch, grad_manual, double_atol=1e-12)
+
+
+@pytest.mark.parametrize("d,N", [(2, 3)])
 def test_branched_sig_backprop_batch(d, N):
     """Batched backprop matches single-path backprop."""
     pysiglib.prepare_branched_sig(d, N)
@@ -646,124 +623,6 @@ def test_branched_sig_backprop_cuda_torch_api(d=2, N=3):
     cpu_grad = path_cpu.grad
 
     check_close(cpu_grad, cuda_grad, double_atol=1e-10)
-
-
-@skip_no_cuda
-def test_branched_sig_combine_torch_api(d=2, N=3):
-    """torch_api branched_sig_combine backward works."""
-    pysiglib.prepare_branched_sig(d, N)
-    np.random.seed(503)
-    bsig1 = torch.tensor(np.random.randn(pysiglib.branched_sig_length(d, N)),
-                         dtype=torch.float64, requires_grad=True)
-    bsig2 = torch.tensor(np.random.randn(pysiglib.branched_sig_length(d, N)),
-                         dtype=torch.float64, requires_grad=True)
-
-    combined = pysiglib.torch_api.branched_sig_combine(bsig1, bsig2, d, N)
-    loss = combined.sum()
-    loss.backward()
-
-    assert bsig1.grad is not None
-    assert bsig2.grad is not None
-    assert bsig1.grad.shape == bsig1.shape
-    assert bsig2.grad.shape == bsig2.shape
-
-
-# ---------------------------------------------------------------------------
-# Combine backprop finite-difference tests
-# ---------------------------------------------------------------------------
-
-@pytest.mark.parametrize("d,N", [(2, 3), (3, 2)])
-def test_branched_sig_combine_backprop_finite_diff(d, N):
-    """Combine backprop matches finite differences."""
-    pysiglib.prepare_branched_sig(d, N)
-    np.random.seed(600)
-    bsig_len = pysiglib.branched_sig_length(d, N)
-    bsig1 = np.random.randn(bsig_len)
-    bsig2 = np.random.randn(bsig_len)
-    derivs = np.random.randn(bsig_len)
-
-    d1, d2 = pysiglib.branched_sig_combine_backprop(derivs, bsig1, bsig2, d, N)
-
-    eps = 1e-8
-    fd1 = np.zeros(bsig_len)
-    for i in range(bsig_len):
-        b1p = bsig1.copy()
-        b1p[i] += eps
-        cp = np.array(pysiglib.branched_sig_combine(b1p, bsig2, d, N))
-        cm = np.array(pysiglib.branched_sig_combine(bsig1, bsig2, d, N))
-        fd1[i] = np.dot(derivs, (cp - cm)) / eps
-    np.testing.assert_allclose(d1, fd1, atol=1e-4)
-
-    fd2 = np.zeros(bsig_len)
-    for i in range(bsig_len):
-        b2p = bsig2.copy()
-        b2p[i] += eps
-        cp = np.array(pysiglib.branched_sig_combine(bsig1, b2p, d, N))
-        cm = np.array(pysiglib.branched_sig_combine(bsig1, bsig2, d, N))
-        fd2[i] = np.dot(derivs, (cp - cm)) / eps
-    np.testing.assert_allclose(d2, fd2, atol=1e-4)
-
-
-# ---------------------------------------------------------------------------
-# CUDA combine tests
-# ---------------------------------------------------------------------------
-
-@skip_no_cuda
-@pytest.mark.parametrize("d,N", [(2, 3)])
-def test_branched_sig_combine_cuda_matches_cpu(d, N):
-    """CUDA combine matches CPU."""
-    pysiglib.prepare_branched_sig(d, N)
-    np.random.seed(601)
-    bsig_len = pysiglib.branched_sig_length(d, N)
-    bsig1 = torch.tensor(np.random.randn(bsig_len), dtype=torch.float64)
-    bsig2 = torch.tensor(np.random.randn(bsig_len), dtype=torch.float64)
-
-    cpu = pysiglib.branched_sig_combine(bsig1, bsig2, d, N)
-    cuda = pysiglib.branched_sig_combine(bsig1.cuda(), bsig2.cuda(), d, N)
-    check_close(cpu, cuda, double_atol=1e-12)
-
-
-@skip_no_cuda
-@pytest.mark.parametrize("d,N", [(2, 3)])
-def test_branched_sig_combine_backprop_cuda_matches_cpu(d, N):
-    """CUDA combine backprop matches CPU."""
-    pysiglib.prepare_branched_sig(d, N)
-    np.random.seed(602)
-    bsig_len = pysiglib.branched_sig_length(d, N)
-    bsig1 = torch.tensor(np.random.randn(bsig_len), dtype=torch.float64)
-    bsig2 = torch.tensor(np.random.randn(bsig_len), dtype=torch.float64)
-    derivs = torch.tensor(np.random.randn(bsig_len), dtype=torch.float64)
-
-    d1_cpu, d2_cpu = pysiglib.branched_sig_combine_backprop(derivs, bsig1, bsig2, d, N)
-    d1_cuda, d2_cuda = pysiglib.branched_sig_combine_backprop(
-        derivs.cuda(), bsig1.cuda(), bsig2.cuda(), d, N)
-    check_close(d1_cpu, d1_cuda, double_atol=1e-10)
-    check_close(d2_cpu, d2_cuda, double_atol=1e-10)
-
-
-@skip_no_cuda
-def test_branched_sig_combine_torch_api_cuda(d=2, N=3):
-    """torch_api branched_sig_combine backward on CUDA."""
-    pysiglib.prepare_branched_sig(d, N)
-    np.random.seed(603)
-    bsig_len = pysiglib.branched_sig_length(d, N)
-    bsig1 = torch.tensor(np.random.randn(bsig_len),
-                         dtype=torch.float64, device='cuda', requires_grad=True)
-    bsig2 = torch.tensor(np.random.randn(bsig_len),
-                         dtype=torch.float64, device='cuda', requires_grad=True)
-
-    combined = pysiglib.torch_api.branched_sig_combine(bsig1, bsig2, d, N)
-    combined.sum().backward()
-
-    assert bsig1.grad is not None
-    assert bsig2.grad is not None
-
-    # Compare against CPU
-    b1_cpu = bsig1.detach().cpu().requires_grad_(True)
-    b2_cpu = bsig2.detach().cpu().requires_grad_(True)
-    pysiglib.torch_api.branched_sig_combine(b1_cpu, b2_cpu, d, N).sum().backward()
-    check_close(b1_cpu.grad, bsig1.grad, double_atol=1e-10)
-    check_close(b2_cpu.grad, bsig2.grad, double_atol=1e-10)
 
 
 # ===========================================================================
@@ -960,26 +819,6 @@ def test_planar_branched_sig_vs_kauri(d, N):
     np.testing.assert_allclose(bsig[1:], ref_reordered, atol=1e-10)
 
 
-@pytest.mark.parametrize("d,N", [(2, 3)])
-def test_planar_branched_sig_concatenation(d, N):
-    """Chen's identity: bsig(X*Y) == branched_sig_combine(bsig(X), bsig(Y))."""
-    pysiglib.prepare_branched_sig(d, N, planar=True)
-    np.random.seed(1002)
-    L1, L2 = 5, 6
-    path1 = np.cumsum(np.random.randn(L1, d) * 0.1, axis=0)
-    path2_increments = np.random.randn(L2 - 1, d) * 0.1
-    path2 = np.vstack([path1[-1:], path1[-1] + np.cumsum(path2_increments, axis=0)])
-
-    bsig1 = pysiglib.branched_sig(path1, N, planar=True)
-    bsig2 = pysiglib.branched_sig(path2, N, planar=True)
-    combined = pysiglib.branched_sig_combine(bsig1, bsig2, d, N, planar=True)
-
-    full_path = np.vstack([path1, path2[1:]])
-    bsig_full = pysiglib.branched_sig(full_path, N, planar=True)
-
-    np.testing.assert_allclose(combined, bsig_full, atol=1e-10)
-
-
 @pytest.mark.parametrize("d,N", [(2, 3), (3, 2)])
 def test_planar_branched_sig_batch(d, N):
     pysiglib.prepare_branched_sig(d, N, planar=True)
@@ -1106,27 +945,6 @@ def test_planar_branched_sig_time_aug_batch(d, N):
     for i in range(B):
         single = pysiglib.branched_sig(paths[i], N, time_aug=True, planar=True)
         np.testing.assert_allclose(batch_result[i], single, atol=1e-12)
-
-
-@pytest.mark.parametrize("d,N", [(2, 3)])
-def test_planar_branched_sig_lead_lag_concatenation(d, N):
-    """Chen's identity should hold with lead_lag enabled for planar sigs."""
-    aug_dim = 2 * d
-    pysiglib.prepare_branched_sig(aug_dim, N, planar=True)
-    np.random.seed(1104)
-    L1, L2 = 5, 6
-    path1 = np.cumsum(np.random.randn(L1, d) * 0.1, axis=0)
-    path2_inc = np.random.randn(L2 - 1, d) * 0.1
-    path2 = np.vstack([path1[-1:], path1[-1] + np.cumsum(path2_inc, axis=0)])
-
-    bsig1 = pysiglib.branched_sig(path1, N, lead_lag=True, planar=True)
-    bsig2 = pysiglib.branched_sig(path2, N, lead_lag=True, planar=True)
-    combined = pysiglib.branched_sig_combine(bsig1, bsig2, aug_dim, N, planar=True)
-
-    full_path = np.vstack([path1, path2[1:]])
-    bsig_full = pysiglib.branched_sig(full_path, N, lead_lag=True, planar=True)
-
-    np.testing.assert_allclose(combined, bsig_full, atol=1e-10)
 
 
 # ---------------------------------------------------------------------------
@@ -1319,69 +1137,3 @@ def test_planar_branched_sig_backprop_cuda_batch(d, N):
         single_grad = pysiglib.branched_sig_backprop(paths[i], bsigs[i], derivs[i], N, planar=True)
         check_close(batch_grad[i], single_grad, double_atol=1e-10)
 
-
-# ---------------------------------------------------------------------------
-# Planar combine backprop tests
-# ---------------------------------------------------------------------------
-
-@pytest.mark.parametrize("d,N", [(2, 3), (3, 2)])
-def test_planar_branched_sig_combine_backprop_finite_diff(d, N):
-    """Planar combine backprop matches finite differences."""
-    pysiglib.prepare_branched_sig(d, N, planar=True)
-    np.random.seed(1500)
-    bsig_len = pysiglib.branched_sig_length(d, N, planar=True)
-    bsig1 = np.random.randn(bsig_len)
-    bsig2 = np.random.randn(bsig_len)
-    derivs = np.random.randn(bsig_len)
-
-    d1, d2 = pysiglib.branched_sig_combine_backprop(derivs, bsig1, bsig2, d, N, planar=True)
-
-    eps = 1e-8
-    fd1 = np.zeros(bsig_len)
-    for i in range(bsig_len):
-        b1p = bsig1.copy()
-        b1p[i] += eps
-        cp = np.array(pysiglib.branched_sig_combine(b1p, bsig2, d, N, planar=True))
-        cm = np.array(pysiglib.branched_sig_combine(bsig1, bsig2, d, N, planar=True))
-        fd1[i] = np.dot(derivs, (cp - cm)) / eps
-    np.testing.assert_allclose(d1, fd1, atol=1e-4)
-
-    fd2 = np.zeros(bsig_len)
-    for i in range(bsig_len):
-        b2p = bsig2.copy()
-        b2p[i] += eps
-        cp = np.array(pysiglib.branched_sig_combine(bsig1, b2p, d, N, planar=True))
-        cm = np.array(pysiglib.branched_sig_combine(bsig1, bsig2, d, N, planar=True))
-        fd2[i] = np.dot(derivs, (cp - cm)) / eps
-    np.testing.assert_allclose(d2, fd2, atol=1e-4)
-
-
-@skip_no_cuda
-@pytest.mark.parametrize("d,N", [(2, 3)])
-def test_planar_branched_sig_combine_cuda_matches_cpu(d, N):
-    pysiglib.prepare_branched_sig(d, N, planar=True)
-    np.random.seed(1501)
-    bsig_len = pysiglib.branched_sig_length(d, N, planar=True)
-    bsig1 = torch.tensor(np.random.randn(bsig_len), dtype=torch.float64)
-    bsig2 = torch.tensor(np.random.randn(bsig_len), dtype=torch.float64)
-
-    cpu = pysiglib.branched_sig_combine(bsig1, bsig2, d, N, planar=True)
-    cuda = pysiglib.branched_sig_combine(bsig1.cuda(), bsig2.cuda(), d, N, planar=True)
-    check_close(cpu, cuda, double_atol=1e-12)
-
-
-@skip_no_cuda
-@pytest.mark.parametrize("d,N", [(2, 3)])
-def test_planar_branched_sig_combine_backprop_cuda_matches_cpu(d, N):
-    pysiglib.prepare_branched_sig(d, N, planar=True)
-    np.random.seed(1502)
-    bsig_len = pysiglib.branched_sig_length(d, N, planar=True)
-    bsig1 = torch.tensor(np.random.randn(bsig_len), dtype=torch.float64)
-    bsig2 = torch.tensor(np.random.randn(bsig_len), dtype=torch.float64)
-    derivs = torch.tensor(np.random.randn(bsig_len), dtype=torch.float64)
-
-    d1_cpu, d2_cpu = pysiglib.branched_sig_combine_backprop(derivs, bsig1, bsig2, d, N, planar=True)
-    d1_cuda, d2_cuda = pysiglib.branched_sig_combine_backprop(
-        derivs.cuda(), bsig1.cuda(), bsig2.cuda(), d, N, planar=True)
-    check_close(d1_cpu, d1_cuda, double_atol=1e-10)
-    check_close(d2_cpu, d2_cuda, double_atol=1e-10)
